Add simple power monitor
The idea of this code is to read the CPU utilization every minute (kind of) and
then at midnight generate a map of the future C states where the CPU will go.
The system will colelct information and every night generate the next day C
states map to be changed at runtime with proper C states framework.

TODO:

Learn: the system will generate a new C states map based on the previuous one.
The system will get the diff perf hour and devide by the number of days he has
been saving ( one array per day ) ,yes we need a linked list :)

After the system is learning we need to make experimetns with phoronix
benchmarks and the % of performance decrement and % of power saving per day
(translate that to money)

Signed-off-by: VictorRodriguez <vm.rod25@gmail.com>